+++ packages/react-table/src/components/Table/Header.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { Header } from './base';
-import { IHeaderRow, TableContext } from './TableContext';
+import { IHeaderRow } from './base';
+import { TableContext } from './TableContext';
 import { ColumnsType } from './base/types';
 
 interface ContextHeaderProps {
