+++ packages/react-table/src/components/Table/enums.ts
@@ -1,0 +1,22 @@
+export enum SortByDirection {
+  asc = 'asc',
+  desc = 'desc'
+}
+
+export enum TableGridBreakpoint {
+  none = '',
+  grid = 'grid',
+  gridMd = 'grid-md',
+  gridLg = 'grid-lg',
+  gridXl = 'grid-xl',
+  grid2xl = 'grid-2xl'
+}
+
+export enum TableVariant {
+  compact = 'compact'
+}
+
+export enum RowSelectVariant {
+  radio = 'radio',
+  checkbox = 'checkbox'
+}
