+++ packages/react-table/src/components/Table/utils/decorators/editable.tsx
@@ -1,13 +1,5 @@
 import * as React from 'react';
-import {
-  IExtra,
-  IFormatterValueType,
-  ITransform,
-  IRowCell,
-  OnRowEdit,
-  RowErrors,
-  RowEditType
-} from '../../TableContext';
+import { IExtra, IFormatterValueType, ITransform, IRowCell, OnRowEdit, RowErrors, RowEditType } from '../../base';
 import { EditColumn } from '../../EditColumn';
 import tableStyles from '@patternfly/react-styles/css/components/Table/table';
