+++ packages/react-table/src/components/Table/utils/decorators/textCenter.ts
@@ -1,2 +1,2 @@
-import { ITransform } from '../../TableContext';
+import { ITransform } from '../../base';
 export const textCenter: ITransform = () => ({ textCenter: true });
